+++ lib/exfmt/syntax_error.ex
@@ -4,6 +4,10 @@ defmodule Exfmt.SyntaxError do
   @type t :: %__MODULE__{line: non_neg_integer, message: String.t}
 
   def exception({line, m1, m2}) do
-    %__MODULE__{line: line, message: m1 <> m2}
+    %__MODULE__{line: line, message: "Error: " <> m1 <> m2}
+  end
+
+  def exception(message, line) do
+    %__MODULE__{line: line, message: "Error: " <> message}
   end
 end